requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
boto3==1.37.9
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.40
langchain-deepseek==0.1.2
langchain-google-genai==2.0.10
langchain-openai==0.3.7
langsmith==0.3.11
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.65.1
orjson==3.10.15
packaging==24.2
phonenumbers==8.13.53
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.10.5
pydantic-extra-types==2.10.2
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
rendercv @ git+https://github.com/code181123/rendercv.git@4a4aa473556fbba8b7a9699cf0120985e7044b8a
rendercv-fonts==0.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
typst==0.12.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
zstandard==0.23.0
flask